+++ datafusion/datasource/src/mod.rs
@@ -35,7 +35,6 @@ use futures::Stream;
 use object_store::{path::Path, ObjectMeta};
 use std::pin::Pin;
 use std::sync::Arc;
-pub mod test;
 
 pub use self::url::ListingTableUrl;
